requirements.txt
Django==1.11.9
djangorestframework==3.7.3
djangorestframework-simplejwt==3.2.1
psycopg2==2.7.3.2
pytz==2018.3
PyJWT==1.5.3
